Moved CGYRO handlers from PORTALS to POWERTORCH
